backend/requirements.txt
google-generativeai
uvicorn
fastapi